[core] Merge GeoJSONTileSource into GeoJSONTileData
