+++ src/mbgl/tile/geojson_tile_data.hpp
@@ -10,8 +10,6 @@ class GeoJSONVT;
 
 namespace mbgl {
 
-class TileSource;
-
 namespace style {
 class UpdateParameters;
 }
@@ -25,9 +23,6 @@ public:
     ~GeoJSONTileData();
 
     void setNecessity(Necessity) final;
-
-private:
-    std::unique_ptr<TileSource> tileSource;
 };
 
 } // namespace mbgl
